// src/app/store/models/employee.model.ts
export class Employee {
    id: string;
    creationTime: Date;
    creatorId: string;
    lastModificationTime: string;
    lastModifierId: string;

    tenantId: string;
    firstName: string;
    middleName: string;
    lastName: string;
    fullName: string;
    photo: string;
    birthPlace: string;
    birthDate: string;
    gender: number;
    religion: number;
    marital: number;
    nationality: number;
    education: number;
    street: string;
    city: string;
    stateProvince: string;
    zipPostal: string;
    phone: string;
    fax: string;
    email: string;
    bankAccount: string;
    nik: string;
    npwp: string;
    address: string;
    userId: string;
    employeeNumber: string;
    employeeStatus: number;
    joinedDate: string;
    terminationDate: string;
    leaveDate: string;
    baseSalary: number;
    salary: number;
    ptkp: number;
    umk: number;
    countryId: string;
    companyId: string;
    departmentId: string;
    jobTitleId: string;
    jobPositionId: string;
    parentId: string;
    deviceId: string;
    userName: string;
    password: string;
    lat: string;
    lng: string;
    radius: number;
    isLocationLockAttendance: boolean;
    isSelfieAttendance: boolean;

    /**
     * 
     * @param model 
     */
    constructor(model?) {
        model = model || {};
        this.id = model.id || null;
        this.creationTime = model.creationTime || null;
        this.creatorId = model.creatorId || null;
        this.lastModificationTime = model.lastModificationTime || null;
        this.lastModifierId = model.lastModifierId || null;

        this.tenantId = model.tenantId || null;
        this.firstName = model.firstName || null;
        this.middleName = model.middleName || null;
        this.lastName = model.lastName || null;
        this.fullName = model.fullName || null;
        this.photo = model.photo || null;
        this.birthPlace = model.birthPlace || null;
        this.birthDate = model.birthDate || null;
        this.gender = model.gender || 0;
        this.religion = model.religion || 0;
        this.marital = model.marital || 0;
        this.nationality = model.nationality || 0;
        this.education = model.education || 0;
        this.street = model.street || null;
        this.city = model.city || null;
        this.stateProvince = model.stateProvince || null;
        this.zipPostal = model.zipPostal || null;
        this.phone = model.phone || null;
        this.fax = model.fax || null;
        this.email = model.email || null;
        this.bankAccount = model.bankAccount || null;
        this.nik = model.nik || null;
        this.npwp = model.npwp || null;
        this.address = model.address || null;
        this.userId = model.userId || null;
        this.employeeNumber = model.employeeNumber || null;
        this.employeeStatus = model.employeeStatus || 0;
        this.joinedDate = model.joinedDate || null;
        this.terminationDate = model.terminationDate || null;
        this.leaveDate = model.leaveDate || null;
        this.baseSalary = model.baseSallary || 0;
        this.salary = model.sallary || 0;
        this.ptkp = model.ptkp || 0;
        this.umk = model.umk || 0;
        this.countryId = model.countryId || null;
        this.companyId = model.companyId || null;
        this.departmentId = model.departmentId || null;
        this.jobTitleId = model.jobTitleId || null;
        this.jobPositionId = model.jobPositionId || null;
        this.parentId = model.parentId || null;
        this.deviceId = model.deviceId || null;
        this.userName = model.userName || null;
        this.password = model.password || null;

        this.lat = model.lat || null;
        this.lng = model.lng || null;
        this.radius = model.radius || 0;
        this.isLocationLockAttendance = model.isLocationLockAttendance || false;
        this.isSelfieAttendance = model.isSelfieAttendance || false;
    }
}